Implement A-(B-(C)-B-(A)-B-(C)) communication model
When humans want to pray to their gods, they transmit pray instead
to clerics. When a human prays for rain, a cleric can give a rain
response on behalf of gods. In the process, a cleric gets a weather
oracle and gets a state of a farm land by invoking human's method.
Then the cleric can determine the rain response by examining farm
land state through the weather oracle.

A stands for Human module,
B stands for Cleric module,
C stands for God module